+++ corehq/apps/domain/urls.py
@@ -177,7 +177,7 @@ domain_settings = [
     url(r'^flags/$', FeatureFlagsView.as_view(), name=FeatureFlagsView.urlname),
     url(r'^toggle_diff/$', toggle_diff, name='toggle_diff'),
     url(r'^sms_rates/$', SMSRatesView.as_view(), name=SMSRatesView.urlname),
-    url(r'^dhis2/', include('corehq.apps.dhis2.urls')),
+    url(r'^dhis2/', include('corehq.motech.dhis2.urls')),
 
     DomainReportDispatcher.url_pattern()
 ]
