+++ corehq/motech/dhis2/static/dhis2/js/dhis2_map_settings.js
