+++ corehq/motech/dhis2/templates/dhis2/log_detail.html
